Success!  I believe I have trimmed down the ClusterMap function and still produce good clustermaps. the PPI plotter now accepts a mask keyword for masking the values.
git-svn-id: svn+ssh://localhost/home/bvr/svnrepos/Programs/RadarClustering@139 c770f1bc-0775-48ed-bd74-6edc8452e29f

+++ LoadRastRadar.py
@@ -1,11 +1,12 @@
 from Scientific.IO.NetCDF import *
+import numpy
 
 def LoadRastRadar(infilename) :
     nc = NetCDFFile(infilename, 'r')
     titleStr = getattr(nc, 'title').replace("Rastified", "Rasterized")
-    lats = nc.variables['lat'].getValue()
-    lons = nc.variables['lon'].getValue()
-    vals = nc.variables['value'].getValue()
+    lats = nc.variables['lat'].getValue().astype(numpy.float32)
+    lons = nc.variables['lon'].getValue().astype(numpy.float32)
+    vals = nc.variables['value'].getValue().astype(numpy.float32)
     timestamp = nc.variables['time'].getValue()[0]
 
     nc.close()
